model.py bugfix: select area
setIndex was called with too big indeces.

setIndex has proven to introduce bugs because it is changing more indeces than one considers.
It did not yet optimize anything, anyway, just duplicated checks.
Therefore I have removed it.

The check for isValidField is not necessary anymore because getNextField always returns a valid field. And if it did not, this would need to be checked for both branches, not just the first.