Shift + Alt + Arrow: moveField
Alt + Arrow (swapField) is useful when you have a wall with a hole and want to move that hole.
For example:  <<<R<<<  =(Alt+Left)=>  <<R<<<<   =(Alt+Left)=>  <R<<<<<
But it is counterintuitive if you want to move an object through a wall:
!!<!!                    !!<!!
!!<!A  =(4x Alt+Left)=>  A!!<!
!!<!!                    !!<!!
(note how the wall is deformed)

Therefore I have introduced a new functionality:
Shift + Alt + Arrow (moveField) moves a field overwriting the object on the destination field.
Before the object is overwritten, however, it is saved in an invisible buffer (_tmpBoard). When the field is moved further the object on the field becoming free is restored from the buffer:
!!<!!                          !!<!!                          !!<!!
!!<!A  =(2x Shift+Alt+Left)=>  !!A!!  =(2x Shift+Alt+Left)=>  A!<!!
!!<!!                          !!<!!                          !!<!!

The buffer is cleared after changes (except for moving fields) or before moving fields other than those which have been moved the last time:
!!<!!                          !!<!!                 !!<!!                          <!!!!
!!<!A  =(2x Shift+Alt+Left)=>  !!A!!  =(Shift+Up)=>  !!A!!  =(2x Shift+Alt+Left)=>  A!!!!
!!<!!                          !!<!!                 !!<!!                          !!<!!

New Algorithm:
	1) save objects on selected fields
	   otherwise when trying to shift a full line from one side to the other one object would get lost.
	   because I am saving all selected fields the order of iteration in the following steps is irrelevant.
	2) save all *un*selected fields
	3) write the values saved in step 1 to their new locations and move the cursors
	4) restore the values of all *un*selected fields (note that these are now other fields than in step 2)