model.py bugfix: swap fields (Alt + Arrow)
The old algorithm had two problems:
(1) if an entire row or column was selected
	it swapped one field too many
(2) if fields were moved across the border
	some objects left the selection because the indeces had the wrong order.
	turned out 690706dc878034d606dd17c9ca4e2d4394a8e58b was not a bugfix but made it worse and the speculation was wrong.

Now, the underlying algorithm is basically still the same, except that it now gets cursors passed instead of indeces and that the ordering has become more complex. If an entire row or column is selected the last field is consciously omitted.