bugfix, prepared settings for getObjectDescription
model.py, Solution.sanityCheck:
objDescr = objects.getObjectDescription(self.model.getField(*cor))

caused exception because defined to take 2 arguments.
the first was a dummy, required so that I could directly use it in gui_image_opener.py:
getLongName = lambda self, obj: objects.getObjectDescription(obj)
there it is saved as an attribute of the getImage object, therefore it is expected to take a self argument.

Now, I have removed the self argument so that the function can be used standalone as one would expect
and have added a lambda function eating up the self argument when passing it to the constructor of ImageOpener.